Extracted zjsunit/namespace.js
